fix(heading): fix stories for heading component
