+++ core/components/atoms/heading/__stories__/Appearance.story.jsx
@@ -1,18 +1,17 @@
 import * as React from 'react';
-import Heading from '../../index';
-import Text from '@/components/atoms/text';
+import { Heading, Text } from '@/index';
 
 // CSF format story
 export const appearance = () => {
   const appearances = ['default', 'subtle', 'disabled', 'white'];
   return (
-    <div className="Row">
+    <div className="d-flex">
       {appearances.map((appear, ind) => {
         return (
-          <div key={ind} className="mr-7">
-            <div className={appear === 'white' ? 'bg-dark' : 'bg-transparent'}>
-              <Heading appearance={appear}>Heading</Heading>
-            </div>
+          <div key={ind} className="d-flex flex-column mr-7">
+            <Heading key={ind} className={appear === 'white' ? 'bg-dark' : 'bg-transparent'} appearance={appear}>
+              Heading
+            </Heading>
             <br />
             <Text weight="strong">{appear.charAt(0).toUpperCase() + appear.slice(1)}</Text>
           </div>
@@ -23,7 +22,7 @@ export const appearance = () => {
 };
 
 export default {
-  title: 'Components/Heading/Variants/Appearance',
+  title: 'Components/Heading/Appearance',
   component: Heading,
   parameters: {
     docs: {
