+++ core/components/atoms/heading/__stories__/style.css
@@ -1,0 +1,3 @@
+.bottom-0 {
+  bottom: 0;
+}
